parse cluster

+++ parse_cluster.py
@@ -21,7 +21,7 @@ def num_redundancies(inp):
 				counts.append([0])
 
 			else: 
-				perc = float(line.split(' ')[-1][:-3])
+				perc = int(line.split(' ')[-1][:-3])
 				print(perc)
